requirements.txt
pandas
numpy
seaborn
matplotlib
scikit-learn
catboost
xgboost
Flask
requests
beautifulsoup4 
gunicorn
dill
python-dotenv
pymango
certifi
pymongo[srv]==3.6
scikit-learn

# -e .